+++ Custom_Swerve/src/main/java/frc/robot/components/SwerveModule.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.components;
 import com.ctre.phoenix6.configs.Slot0Configs;
 import com.ctre.phoenix6.controls.PositionVoltage;
 import com.ctre.phoenix6.hardware.CANcoder;
@@ -16,7 +16,7 @@ public class SwerveModule {
     private double offset;
     private double direction;
     Slot0Configs pidConfigs = new Slot0Configs();
-    SwerveModule(TalonFX driveMotor, TalonFX turnMotor, CANcoder coder, double offset, double direction){
+    public SwerveModule(TalonFX driveMotor, TalonFX turnMotor, CANcoder coder, double offset, double direction){
         this.driveMotor = driveMotor;
         this.turnMotor = turnMotor;
         this.coder = coder;
